further clean-up and user input for experiments
